requirements.txt
ansicolors==1.1.8
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
grpcio==1.76.0
grpcio-reflection==1.76.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
labgrid==25.0.1
MarkupSafe==3.0.3
packaging==26.0
pexpect==4.9.0
pluggy==1.6.0
protobuf==6.33.4
ptyprocess==0.7.0
Pygments==2.19.2
pyserial-labgrid==3.5.0.2
pytest==9.0.2
pyudev==0.24.4
pyusb==1.3.1
PyYAML==6.0.3
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
xmodem==0.4.7
